+++ module/Application/src/Module.php
@@ -2,10 +2,16 @@
 
 namespace Application;
 
+use Zend\EventManager\EventInterface;
 use Zend\ModuleManager\Feature\ConfigProviderInterface;
+use Zend\ModuleManager\Feature\BootstrapListenerInterface;
+use Zend\Log\Logger;
+use Zend\Log\Writer\Stream;
 
-class Module implements ConfigProviderInterface
+class Module implements ConfigProviderInterface, BootstrapListenerInterface
 {
+    private const ERROR_LOG_FILE = __DIR__ . '/../logs/errors.log';
+
     public function getConfig(): array
     {
         return array_merge(
@@ -18,5 +24,24 @@ class Module implements ConfigProviderInterface
             require __DIR__ . '/../config/rbac.config.php',
             require __DIR__ . '/../config/twig.config.php'
         );
+    }
+
+    public function onBootstrap(EventInterface $e)
+    {
+        /** @var \Zend\Mvc\MvcEvent $e */
+        $sharedManager = $e->getApplication()->getEventManager()->getSharedManager();
+        $logger = $this->getLogger();
+        $sharedManager->attach('Zend\Mvc\Application', 'dispatch.error', function($e) use ($logger) {
+            if ($e->getParam('exception')) {
+                $logger->crit($e->getParam('exception'));
+            }
+        });
+    }
+
+    protected function getLogger(): Logger
+    {
+        $logger = new Logger;
+        $writer = new Stream(self::ERROR_LOG_FILE);
+        return $logger->addWriter($writer);
     }
 }