+++ module/Cli/src/Manager/Provider/CommandManagerProviderInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Cli\Manager\Provider;
+
+interface CommandManagerProviderInterface
+{
+    public function getCommandManager(): array;
+}