+++ module/Cli/src/Options/ModuleOptionsFactory.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Cli\Options;
+
+use Interop\Container\ContainerInterface;
+use Zend\ServiceManager\Factory\FactoryInterface;
+
+class ModuleOptionsFactory implements FactoryInterface
+{
+    public function __invoke(ContainerInterface $container, $requestedName, array $options = null): ModuleOptions
+    {
+        return new ModuleOptions($container->get('config')['cli']);
+    }
+}