+++ module/ExAssetic/config/cli.config.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace ExAssetic;
+
+return [
+    'cli' => [
+        'commands' => [
+            'ExAssetic\Command\BuildCommand',
+            'ExAssetic\Command\SetupCommand',
+        ],
+    ],
+    'command_manager' => [
+        'factories' => [
+            'ExAssetic\Command\BuildCommand' => Command\BuildCommandFactory::class,
+            'ExAssetic\Command\SetupCommand' => Command\SetupCommandFactory::class,
+        ],
+    ],
+];