+++ module/ExAssetic/src/Module.php
@@ -16,7 +16,8 @@ class Module implements ConfigProviderInterface
     {
         return array_merge(
             require __DIR__ . '/../config/module.config.php',
-            require __DIR__ . '/../config/rbac.config.php'
+            require __DIR__ . '/../config/rbac.config.php',
+            require __DIR__ . '/../config/cli.config.php'
         );
     }
 }