+++ module/User/src/Service/SignUpServiceFactory.php
@@ -10,7 +10,7 @@ class SignUpServiceFactory implements FactoryInterface
     public function __invoke(ContainerInterface $container, $requestedName, array $options = null): SignUpService
     {
         $userRepository = $container->get('Doctrine\ORM\EntityManager')->getRepository('User\Entity\User');
-        $passwordHashService = $container->get('Di')->get('Zend\Crypt\Password\BcryptSha');
+        $passwordHashService = $container->get('User\Crypt\PasswordCrypt');
         return new SignUpService($userRepository, $passwordHashService);
     }
 }